test: Move columns tests from e2e to integration layer.
